Unify server setup scripts [1] #789 #767  shell script needs to support productive and development systems (setting on top of file)  remove batch script and add explanation on how to execute the shell script on windows systems  rename to server-setup.sh  increase sormas domain java heap size to 1024 (-Xmx1024) see #232  glassfish.sh edit line 58 (also in sormas_base)  Remove unnecessary "Press enter" in glassfish.sh  Add note for need of bash-software to execute glassfish.sh

+++ sormas-base/setup/server-setup.sh
@@ -1,12 +1,15 @@
 #  Requirements:
-#  * Payara 4.1.1 installed and proper Java-Version configured in asenv.conf 
+#  * Payara 4.1.2.172 installed and proper Java-Version configured in asenv.conf 
 #  * Database created
-#  * glassfish-config-directory prepared
-#  * directory /var/log/glassfish/${DOMAIN_NAME} exists and is owned by user 'glassfish'
-
+#  * setup directory prepared
+#  * log directory (e.g. /var/log/glassfish/${DOMAIN_NAME}) exists and is owned by user 'glassfish'
 
 
 # ------ Config BEGIN ------
+
+# DEVELOPMENT ENVIRONMENT OR PRODUCTIVE SERVER?
+#DEV_SYSTEM=true
+DEV_SYSTEM=false
 
 # GLASSFISH
 GLASSFISH_HOME=/opt/payara-172/glassfish
@@ -14,6 +17,8 @@ DOMAIN_NAME=sormas
 DOMAINS_HOME=/opt/domains
 DOMAIN_DIR=${DOMAINS_HOME}/${DOMAIN_NAME}
 LOG_HOME=/var/log/glassfish/sormas
+TEMP_DIR=/opt/sormas-temp
+GENERATED_DIR=/opt/sormas-generated
 PORT_BASE=6000
 PORT_ADMIN=6048
 
@@ -28,7 +33,7 @@ DB_SERVER=localhost
 DB_PORT=5432
 
 # MAIL
-MAIL_FROM=noreply@symeda.de
+MAIL_FROM=
 
 # ------ Config END ------
 
@@ -43,32 +48,31 @@ echo "Admin Port: ${PORT_ADMIN}"
 
 read -p "Press [Enter] to continue..."
 
-#setting owner of the deploy-directory to glassfish, in order to maintain proper permissions
-#chown -R glassfish:glassfish /root/deploy/
-
-# patching gf-modules
-##removing old versions
-#rm ${GLASSFISH_HOME}/modules/jboss-logging.jar
-
-##placing new versions
-#cp gf-modules/*.jar ${GLASSFISH_HOME}/modules/
+# create needed directories and set user rights
+mkdir -R ${TEMP_DIR}
+mkdir -R ${GENERATED_DIR}
+setfacl -m u:postgres:rwx ${TEMP_DIR} 
+setfacl -m u:glassfish:rwx ${TEMP_DIR}
+setfacl -m u:postgres:rwx ${GENERATED_DIR}
+setfacl -m u:glassfish:rwx ${GENERATED_DIR}
 
 # setting ASADMIN_CALL and creating domain
 ASADMIN="${GLASSFISH_HOME}/bin/asadmin --port ${PORT_ADMIN}"
-${GLASSFISH_HOME}/bin/asadmin create-domain --domaindir ${DOMAINS_HOME} --portbase ${PORT_BASE} ${DOMAIN_NAME}
+${GLASSFISH_HOME}/bin/asadmin create-domain --domaindir ${DOMAINS_HOME} --portbase ${PORT_BASE} --nopassword ${DOMAIN_NAME}
 
 read -p "Press [Enter] to continue..."
 
-# copying server-libs
-cp serverlibs/*.jar ${DOMAIN_DIR}/lib/
+if [ $DEV_SYSTEM != true ] then
+  # copying server-libs
+  cp serverlibs/*.jar ${DOMAIN_DIR}/lib/
 
-# copying bundles
-mkdir -p ${DOMAIN_DIR}/autodeploy/bundles
-cp -a bundles/*.jar ${DOMAIN_DIR}/autodeploy/bundles/
+  # copying bundles
+  mkdir -p ${DOMAIN_DIR}/autodeploy/bundles
+  cp -a bundles/*.jar ${DOMAIN_DIR}/autodeploy/bundles/
 
-echo "copying libs completed"
-read -p "Press [Enter] to continue..."
-
+  echo "copying libs completed"
+  read -p "Press [Enter] to continue..."
+fi
 
 cat << END > ${DOMAIN_DIR}/config/login.conf
 ${DOMAIN_NAME}Realm { org.wamblee.glassfish.auth.FlexibleJdbcLoginModule required; };
@@ -79,6 +83,9 @@ read -p "Press [Enter] to continue..."
 
 ${GLASSFISH_HOME}/bin/asadmin start-domain --domaindir ${DOMAINS_HOME} ${DOMAIN_NAME}
 read -p "Press [Enter] to continue..."
+
+# General domain settings
+${ASADMIN} create-jvm-options -Xmx1024m
 
 # JDBC pool
 ${ASADMIN} create-jdbc-connection-pool --restype javax.sql.ConnectionPoolDataSource --datasourceclassname org.postgresql.ds.PGConnectionPoolDataSource --isconnectvalidatereq true --validationmethod custom-validation --validationclassname org.glassfish.api.jdbc.validation.PostgresConnectionValidation --property "portNumber=${DB_PORT}:databaseName=${DB_NAME}:serverName=${DB_SERVER}:user=${DB_USER}:password=${DB_PW}" ${DOMAIN_NAME}DataPool
@@ -118,49 +125,30 @@ ${ASADMIN} set-log-attributes com.sun.enterprise.server.logging.GFFileHandler.fi
 ${ASADMIN} set-log-attributes com.sun.enterprise.server.logging.GFFileHandler.maxHistoryFiles=14
 ${ASADMIN} set-log-attributes com.sun.enterprise.server.logging.GFFileHandler.rotationLimitInBytes=0
 ${ASADMIN} set-log-attributes com.sun.enterprise.server.logging.GFFileHandler.rotationOnDateChange=true
-${ASADMIN} set-log-levels org.wamblee.glassfish.auth.HexEncoder.level=SEVERE
-${ASADMIN} set-log-levels javax.enterprise.system.util.level=SEVERE
+${ASADMIN} set-log-levels org.wamblee.glassfish.auth.HexEncoder=SEVERE
+${ASADMIN} set-log-levels javax.enterprise.system.util=SEVERE
 
 read -p "Press [Enter] to continue..."
 
-#Login-Auditierung aktivieren
-#${ASADMIN} set configs.config.server-config.security-service.audit-enabled=true
-#${ASADMIN} create-audit-module --classname=de.symeda.glassfish.audit.LoginAttemptAuditModule --target=server-config LoginAttemptAudit
+if [ $DEV_SYSTEM != true ] then
+  #make the glassfish listen to localhost only
+  ${ASADMIN} set configs.config.server-config.http-service.virtual-server.server.network-listeners=http-listener-1
+  ${ASADMIN} delete-network-listener --target=server-config http-listener-2
+  ${ASADMIN} set configs.config.server-config.network-config.network-listeners.network-listener.admin-listener.address=127.0.0.1
+  ${ASADMIN} set configs.config.server-config.network-config.network-listeners.network-listener.http-listener-1.address=127.0.0.1
+  ${ASADMIN} set configs.config.server-config.iiop-service.iiop-listener.orb-listener-1.address=127.0.0.1
+  ${ASADMIN} set configs.config.server-config.iiop-service.iiop-listener.SSL.address=127.0.0.1
+  ${ASADMIN} set configs.config.server-config.iiop-service.iiop-listener.SSL_MUTUALAUTH.address=127.0.0.1
+  ${ASADMIN} set configs.config.server-config.jms-service.jms-host.default_JMS_host.host=127.0.0.1
+  ${ASADMIN} set configs.config.server-config.admin-service.jmx-connector.system.address=127.0.0.1
 
-read -p "Press [Enter] to continue..."
-
-#make the glassfish listen to localhost only
-${ASADMIN} set configs.config.server-config.http-service.virtual-server.server.network-listeners=http-listener-1
-${ASADMIN} delete-network-listener --target=server-config http-listener-2
-${ASADMIN} set configs.config.server-config.network-config.network-listeners.network-listener.admin-listener.address=127.0.0.1
-${ASADMIN} set configs.config.server-config.network-config.network-listeners.network-listener.http-listener-1.address=127.0.0.1
-${ASADMIN} set configs.config.server-config.iiop-service.iiop-listener.orb-listener-1.address=127.0.0.1
-${ASADMIN} set configs.config.server-config.iiop-service.iiop-listener.SSL.address=127.0.0.1
-${ASADMIN} set configs.config.server-config.iiop-service.iiop-listener.SSL_MUTUALAUTH.address=127.0.0.1
-${ASADMIN} set configs.config.server-config.jms-service.jms-host.default_JMS_host.host=127.0.0.1
-${ASADMIN} set configs.config.server-config.admin-service.jmx-connector.system.address=127.0.0.1
-
-read -p "Press [Enter] to continue..."
-
-#Applications deployen
-#cp -a applications/*.*ar ${DOMAIN_DIR}/autodeploy/
-
-#read -p "Press [Enter] to continue..."
-
-
-#templates einfügen
-#mkdir ${DOMAIN_DIR}/templates/
-#cp -a templates/* ${DOMAIN_DIR}/templates/
-
-#read -p "Press [Enter] to continue..."
+  read -p "Press [Enter] to continue..."
+fi
 
 
 ${GLASSFISH_HOME}/bin/asadmin stop-domain --domaindir ${DOMAINS_HOME} ${DOMAIN_NAME}
 
 read -p "Press [Enter] to continue..."
-
-##do not run the server here (as in bat script), because it will be executed as root
-#${GLASSFISH_HOME}/bin/asadmin start-domain --domaindir ${DOMAINS_HOME} ${DOMAIN_NAME}
 
 chown -R glassfish:glassfish ${GLASSFISH_HOME}
 
@@ -174,4 +162,3 @@ echo '  - JVM parameters fit?'
 echo '  - payara-default-domains deleted?'
 echo '  - Java-Version in asenv.conf?'
 echo '  - Apache properly configured?'
-echo '  - using configuration from an older domain (properties, logback)?'
